Allow to pass all the relevant info to QQuickPixmap::isCached
If we don't allow to pass the QQuickImageProviderOptions to the isCached
we are missing one of the three components that identify if the pixmap is actually cached.

Old code is ok for the QQuickAnimatedImagePrivate use case since it always creates
QQuickPixmap with the default QQuickImageProviderOptions options, but if someone else
wants to use QQuickPixmap::isCached they will need to be able to pass
the options part of the cache key.

Change-Id: I3153ad9ed30e7332b5cb4896dcebd408f2bd9afe
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>