QQuickWindow: remove sendFilteredTouchEvent
The functionality is now completely assimilated into
sendFilteredPointerEvent.

Flickable can now steal the touch grab from a TapHandler child,
for example: the TapHandler grabbed the eventpoint, whereas
QQuickFlickable::childMouseEventFilter() was given a synthetic
mouse event derived from the touchpoint which the eventpoint
represented, and we fully follow through the consequences of that.
If an Item filters a press event, avoid doing normal delivery
to that same Item again.

Change-Id: Icd9c88ab752f2b728f7d612504013c6dc72ff9fe
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>