Fix compilation of qmldevtools on Windows when cross-compiling
The use of wildcards in SOURCES is an unspecified "feature" and
it turns out that it doesn't work when doing option(host_build)
and cross-compiling. We could use $$files(), but there is no
truly good reason for using it for only some directories and not
others, so this is now a manual expansion.

Task-number: QTBUG-58705
Change-Id: Ib3bd7bbba01fe4130703d5635fa344b62caea5bb
Reviewed-by: Lars Knoll <lars.knoll@qt.io>