Fix a bug in Chunk::sortIntoBins()
The bug lead to us not sorting all freed memory into
the bins for the allocator to find. The lead to memory
being 'lost' in the garbage collector.

Add an assertion that checks for lost memory when
we're running with aggressive GC.

Also make sure we don't run GC twice in a row when
aggressive GC is enabled.

Change-Id: I4fb6732acce8a2e66258fa70fb7d8f1f939cfd9f
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>