Shaders: Fix crash when disconnecting shaders from item properties
The disconnect was done with a signal index local to the item, instead
of 'global' to the item and all its parents. The index also had to be a
method index, not a signal index. Together this resulted in a failing
disconnect, which would keep SlotObjects alive with outdated data. Any
subsequent use would result in a crash.

Task-number: QTBUG-58336
Change-Id: I26b1868b2f8c61a4ffab1c72c85178632ca87599
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>