notify a PointerHandler when it loses grab due to Item::grabTouchPoints
and move more notification responsibility into QQuickEventPoint,
thus simplifying QQuickWindowPrivate::grabTouchPoints() which is the
implementation behind QQuickItem::grabTouchPoints.
It's important for QQuickEventPoint::setGrabberItem to
change local state first and then notify, to prevent recursive
notify/ungrab loops.  MPTA for example does an ungrab
when it receives touchUngrabEvent, which then notifies again
if the first ungrab was not already fully completed.

Change-Id: I6f7b939c8cd76ac5f2d1ddda8b210fa3d31d619a
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>