QQuickImageBase: allow derived classes to affect devicePixelRatio usage
QQuickImageBase currently sets its d->devicePixelRatio member in three
scenarios:

- The URL uses the image: scheme (an image provider)
- The URL is an svg/svgz
- The URL contains the "@2x" syntax

In all other cases, it sets d->devicePixelRatio to 1.0. QQuickIconImage
derives from QQuickImageBase, and can display theme icons in addition
to regular images. Theme icons have their own directory structure that
determines what is and is not a high DPI image:

https://standards.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html

In particular, they use e.g. a 22x22@2 directory to indicate that
the icons contained within it are high DPI images, as opposed to
storing that information in the file name.

To account for these images, QQuickIconImage must be able to affect
the decision making progress that determines whether or not an image
is loaded at a higher DPI. This patch adds a virtual function to
QQuickImageBasePrivate that allows subclasses to have some control over
this.

Change-Id: Id45e0580c8e2ea9a96a41c54ef5fa765ce5922cf
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>