ApplicationWindow: fix typo
It was meant to be called contentData, as the getter is called,
but the part in the Q_PROPERTY macro that specifies the name of
the property contained "data" in place of "contentData".

[ChangeLog][Controls][ApplicationWindow] The data-property has
been renamed to contentData, as it was documented. Notice that
a data-property still exists in QML Window, but is no longer
overshadowed in ApplicationWindow.

Task-number: QTBUG-58932
Change-Id: I5251d42f5ae5ef6d04d87c3f8d20d7adbefe1eb5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>