QQuickLayout: ensure that all change listeners are removed
The first fix for QTBUG-51927 (59c6c0e0) went to 5.6.2, where only the
SiblingOrder ChangeType was listened to by QQuickLayout. It was then
cherry-picked to 5.7 (5149aa68), where SiblingOrder was still the only
ChangeType in use.

3b4f00ec then optimized QQuickLayout by swapping connections for item
change listeners, but didn't check for usages of change types elsewhere
in the file.

This patch moves the change types into a variable that ensures there
is one place for future changes. 5.6 is not affected.

Task-number: QTBUG-51927
Change-Id: Ifd6e0545ce543ab79d6415e007b35c457cacc83a
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>