QQuickStackView: fix pop/replace exit leak
QQuickStackView did not keep track of the element that was exiting
the view during a pop/replace exit transition, but it was expecting
viewItemTransitionFinished() to be called and cleaned up the element
there. If a pop/replace exit transition was running at the destruction
time, QQuickStackView did not clean up the element at all (because it
had been already removed from the internal stack of elements) and thus
the transition was left running.

Task-number: QTBUG-59034
Change-Id: I3cd7c761fab79574b9a7cc849788f36765def45b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>