Fix QQuickStackViewPrivate::startTransition()
This seems like a mistake that slipped in when popTransition(),
pushTransition() and replaceTransition() were merged in acecb06.
Only immediate push/pop/replace operations were affected, where
the animations were force-completed by hand. In this scenario it
was always passing the pop transitions for completion, even if
the current operation was push or replace.

Change-Id: I6a940953feaefe6643ddae26c070856a66bf6c10
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>