Software: Flush whole window on Expose event
The software render loop updates the window content via either an expose
event or an update event.  An expose event should always flush the entire
backingstore, but previously only the newly updated region would be
flushed. The first time a window is exposed this is fine because the whole
scene would be rendered, but when hidden and shown again, the software
renderer might show that nothing has changed, and nothing would be
flushed. A new flag has been added that forces a full window flush for an
expose event.

Task-number: QTBUG-59177
Change-Id: I3700bbcc76bc97be4eb0c822e2945ebef339f11a
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>