Fix cache file generation for prefix and shadow builds
For prefix builds generate the cache files in $$MODULE_BASE_OUTDIR/qml/
and install them from there.

The use of relative paths for the qml cache extra compiler output is
required because target_predeps in the extra compiler configuration will
generate relative paths as dependency for the target.

Task-number: QTBUG-58570
Change-Id: I6eedfd2aca1b0bdc7a230ce7521e499c6ab70ee0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>