Wearable: remove unnecessary wrapper items
Almost all pages were centering a fixed-size SwipeView. If the window
was larger than assumed, swipes would only work in the center are of
the window. Therefore it's better to stretch the view, and just center
the content instead. This way swipes work consistently in the whole
content area of the page.

Change-Id: Idc2a2bc327a49314cbf1271fa34b74df5f633ed8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>