QSGDistanceFieldUtil: Remove a layer of indirection
These mysterious setters appear to be unused (and always were, as far
as I can see). How useful they are is not entirely clear (as opposed to
a patch to the source for instance, and as it's private API anyway, you'd be
touching internals to use this. Additionally, removing the indirection makes
the code a bit cleaner and more self-contained in my opinion.

This removal leaves the value of QSGDistanceFieldGlyphCacheManager in
some question to me, given that it's basically a glorified wrapper
around QHash.

Change-Id: I6d18eb40d8cd00ebe389b4ed53448f3401962ae6
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>
Reviewed-by: Yoann Lopes <yoann.lopes@qt.io>