Simplify internal source access API
All call sites of QQmlDataBlob::Data immediately convert the data to
UTf-8 - for .qml, .js and qmldir files. We can simplify the code by
reflecting that in the API and that also opens up the possibility for
future optimizations.

This means that the bi-pointer has to go, but at the moment the Data
object is stack-allocated anyway.

Since the "Data" class always represents source code, it is now called
SourceCodeData.

Change-Id: Icd262ed1e35f9edd64945ba6c16d80f9917eae72
Reviewed-by: Lars Knoll <lars.knoll@qt.io>