QQuickComboBox: don't emit highlightedIndexChanged() during destruction
If QQuickComboBox was destroyed while its popup was open, it emitted
highlightedIndexChanged() during the destruction due to the popup's
visibility change. Use setPopup(nullptr) to destroy the popup in a
controlled manner, that is, disconnecting the relevant signals first.
This way deleting the popup doesn't have weird side-effects during
the destruction of QQuickComboBox.

Task-number: QTBUG-57650
Change-Id: I1c55707be68a24e115276ffb6dfa1c89051a45ea
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>