Make scaled image size calculation available to image providers
When reduced size image loading is requested through
Image::sourceSize, the precise calculation of the scaled size to load
is somewhat complex. This commit moves that calculation into a function
accessible to image providers, so that they may use the same scaling
as Quick's default image loading, and avoid code duplication.

This is a code restructuring only; no behavior should change.

Change-Id: Ic965fb2b6c22a5d21add41b8395e3a3842697d20
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>