Flickable: don't steal grab from item with keepTouchGrab set
Sometimes filterMouseEvent can be given a mouse event which was
synthesized from a touch point which was already grabbed.  In such
cases, if the grabber has keepTouchGrab set, respect it.
This makes it possible again for a QQControls 2 slider to be draggable
inside a Flickable which is also draggable in the same direction.

Task-number: QTBUG-59416
Change-Id: I93b7fd9cb846b1e574615154f9a54316b60c0477
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>