add pinchDragFlingMPTA example
This is a prototype of what 3-finger pinch is intended for:
manipulating a container exclusively with a 3-finger gesture while
not interfering at all with other gestures that occur inside the
container.  And it's normal that existing Items (like MPTA) do
implicit grabs of touchpoints just by accepting, so this is also
a test of cooperation between PointerHandlers and legacy Areas.
It also adds the ability to drag the MPTA only while the Meta
key is held down.  And there is momentum on release whether you
are doing a 3-finger pinch or via the DragHandler.

Change-Id: Icd0e84809ec32dc8f347dd9c8f875d10f52eba19
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>