Start over with event delivery when touchpoint presses occur
It was part of the intended design for pointer handlers; omitting this
feature has until now prevented some opportunities for handoff
from one handler to another.

Now if PinchHandler has grabbed, pressing one more point will let
something else have a chance again, for example.

But this is incomplete in that if we release a point, we should do
the same thing, actually.

Change-Id: I10f567e7e4388bf0caab54c261178f19db20b14a
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>